+++ app/models/product.rb
@@ -1,5 +1,6 @@
 class Product < ApplicationRecord
   include PgSearch::Model
+  include Favoritable
   pg_search_scope :search_full_text, against: {
     title: 'A',
     description: 'B'
@@ -17,21 +18,8 @@ class Product < ApplicationRecord
   validates :price, presence: true
   belongs_to :category
   belongs_to :user, default: -> { Current.user }
-  has_many :favorites, dependent: :destroy
 
   def owner?
     user_id == Current.user&.id
-  end
-
-  def favorite!
-    favorites.create(user: Current.user)
-  end
-
-  def unfavorite!
-    favorite.destroy
-  end
-
-  def favorite
-    favorites.find_by(user: Current.user)
   end
 end
